requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
openai==1.3.5
python-dotenv==1.0.0
aiofiles==23.2.1

